Taste The Rainbow

Replacement for cat, message, and print.

lolcat(..., file = "", fill = FALSE, lol = NULL)

lolmessage(..., domain = NULL, appendLF = TRUE, lol = NULL)

lolprint(x, ..., lol = NULL)

Arguments

...
Arguments as for cat, including sep, etc, arguments passed to message or extra arguments to print

file
Used to check that we don't put silly output into actual files. Base cat will be used when file is given

fill
Passed to cat

lol
Optional lol engine to use. If omitted we use the default one registered when the package is loaded

domain
used by translations (not yet supported)

appendLF
logical: should messages given as a character string have a newline appended? (see message)

x
Object to print

Examples

for (i in 1:20) {
  lolcat("hello world\n")
}

#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world
#> hello world